# SCons: Disable C++ exception handling