Upon investigating the extremely slow MSVC build times in #80513, I noticed that while Godot policy is to never use exceptions, we weren't enforcing it with compiler flags, and thus still included exception handling code and stack unwinding.

This is wasteful on multiple aspects:

- Binary size: Around 20% binary size reduction with exceptions disabled for both MSVC and GCC binaries.
- Compile time:
  * More than 50% build time reduction with MSVC.
  * 10% to 25% build time reduction with GCC + LTO.
- Performance: Possibly, needs to be benchmarked.

Since users may want to re-enable exceptions in their own thirdparty code or the libraries they compile with Godot, this behavior can be toggled with the `disable_exceptions` SCons option, which defaults to true.

Alternative to #80516.
Fixes #80513.

Could be considered for a `4.1` cherry-pick, but this will require significant testing to make sure it behaves as expected, and it may be a welcome but surprising change for existing 4.1.x users, so we might opt for not including it.

----

## Test builds

**Edit:** New test builds for all platforms made with the official buildsystem: https://downloads.tuxfamily.org/godotengine/testing/4.2-dev-pr80612/

To help validate this PR, I've done some builds of the before and after states (from this PR, so on top of 0308422f461dce11339896249e23ff02d978bfa0), for Linux (GCC on our official buildsystem) and Windows (VS 2022 on my laptop, unsigned).

You'll find zips here:
https://downloads.tuxfamily.org/godotengine/testing/4.2.dev.no-exceptions/

#### Linux builds

```
-rwxr-xr-x. 1 akien akien 116M Aug 14 13:11 godot.linuxbsd.editor.x86_64.exceptions-with-oidn
-rwxr-xr-x. 1 akien akien 110M Aug 14 11:02 godot.linuxbsd.editor.x86_64.exceptions-without-oidn
-rwxr-xr-x. 1 akien akien  97M Aug 14 13:11 godot.linuxbsd.editor.x86_64.no-exceptions-with-oidn
-rwxr-xr-x. 1 akien akien  90M Aug 14 11:02 godot.linuxbsd.editor.x86_64.no-exceptions-without-oidn
-rwxr-xr-x. 1 akien akien  68M Aug 14 13:11 godot.linuxbsd.template_release.x86_64.exceptions
-rwxr-xr-x. 1 akien akien  57M Aug 14 13:11 godot.linuxbsd.template_release.x86_64.no-exceptions
```

`exceptions` are the same as we currently use for official builds, and `no-exceptions` use the new option to disable exception handling.

OIDN is special cased as I couldn't disable exception handling for it yet, so the `with-oidn` builds actually still have exceptions enabled for OIDN's build environment only. `without-oidn` builds were made with `module_denoise_enabled=no`. It's an editor-only dependency so it doesn't affect the template build.

I didn't include the `without-oidn` builds in the above ZIPs to save your bandwidth, I mostly build them to see the impact on size and build time.

These builds were made on the official buildsystem with the same options as official releases, so with full optimizations and LTO enabled (`scons p=linuxbsd target=editor warnings=no production=yes`)

#### Windows builds

```
-rwxr-xr-x 1 Remi 197121 118M Aug 14 11:17 godot.windows.editor.x86_64.exceptions.exe*
-rwxr-xr-x 1 Remi 197121 101M Aug 14 13:50 godot.windows.editor.x86_64.no-exceptions.exe*
-rwxr-xr-x 1 Remi 197121  54M Aug 14 12:17 godot.windows.template_release.x86_64.exceptions.exe*
-rwxr-xr-x 1 Remi 197121  51M Aug 14 11:52 godot.windows.template_release.x86_64.no-exceptions.exe*
```

## Impact on size

As seen above, for Linux builds with GCC 10, we get a 16% size reduction for the editor build (or said otherwise, enabling exceptions added 20% size), and a similar ratio for the release template.

With LTO disabled, both build types are around 3M bigger, so the ratio doesn't change much.

Dev build: 758M to 736M on Linux, so 3% size reduction. Less than I expected.

On Windows, there's a similar 15% size reduction for the editor build, and only 5% size reduction for the release template (but it was already significantly smaller than the Linux one, now they have a comparable size).

## Impact on compile time

#### Linux with GCC 10

**Production builds with LTO enabled.**

`scons p=linuxbsd target=editor warnings=no production=yes`

```
==> ../logs/godot-editor-exceptions-with-oidn.log <==
[Time elapsed: 00:09:19.461]

==> ../logs/godot-editor-exceptions-without-oidn.log <==
[Time elapsed: 00:09:09.282]

==> ../logs/godot-editor-no-exceptions-with-oidn.log <==
[Time elapsed: 00:08:51.862]

==> ../logs/godot-editor-no-exceptions-without-oidn.log <==
[Time elapsed: 00:07:55.473]

==> ../logs/godot-template_release-exceptions.log <==
[Time elapsed: 00:06:10.944]

==> ../logs/godot-template_release-no-exceptions.log <==
[Time elapsed: 00:04:34.993]
```

The differences aren't huge, because LTO still takes a while, and it seems like having to compile OIDN with exceptions enabled adds a whole minute to the build time.

If we compare the builds without OIDN, there's a 13% build time reduction for the editor, and a 26% build time reduction for the release template.

**Builds without LTO enabled.**

`no-lto` is `scons p=linuxbsd target=editor warnings=no`, while `dev-build` is `scons p=linuxbsd target=editor dev_build=yes`.

```
==> ../ext/logs/godot-editor-dev-build-exceptions.log <==
[Time elapsed: 00:05:47.942]

==> ../ext/logs/godot-editor-dev-build-exceptions.log <==
[Time elapsed: 00:06:00.398]

==> ../ext/logs/godot-editor-no-lto-exceptions.log <==
[Time elapsed: 00:10:02.207]

==> ../ext/logs/godot-editor-no-lto-no-exceptions.log <==
[Time elapsed: 00:07:35.507]
```

For an optimized build without LTO (so not paying a disproportionate amount of time for linking), there's a significant 25% build time speedup.

For the dev builds, I actually got a 2% build time increase, though I don't know if it's representative, I ran the test only once.

Sizes for these builds for those interested:
```
-rwxr-xr-x. 1 root root 758M Aug 14 11:45 godot.linuxbsd.editor.dev.x86_64.exceptions-dev-build
-rwxr-xr-x. 1 root root 736M Aug 14 12:01 godot.linuxbsd.editor.dev.x86_64.no-exceptions-dev-build
-rwxr-xr-x. 1 root root 119M Aug 14 11:34 godot.linuxbsd.editor.x86_64.exceptions-no-lto
-rwxr-xr-x. 1 root root 100M Aug 14 11:19 godot.linuxbsd.editor.x86_64.no-exceptions-no-lto
```

#### Windows with MSVC from VS 2022

```
scons p=windows target=editor scu_build=yes verbose=yes

- exceptions
[Time elapsed: 00:34:24.411]

- no-exceptions
[Time elapsed: 00:12:18.504]


scons p=windows target=template_release scu_build=yes verbose=yes

- exceptions
[Time elapsed: 00:19:28.350]

- no-exceptions
[Time elapsed: 00:09:20.439]
```

I used `scu_build` to speed things up, so this exacerbates issues like #80513 since the parallel build goes faster, but we'll still get stuck a long time on `gdscript_vm.cpp` for the `exceptions` build.

The results are pretty impressive though, with the build time divided by 3 for the editor (34m to 12m) and divided by 2 for the template (19m to 9m).

## Impact on performance

Help welcome to run some benchmarks! You can use the builds linked above.

## Impact on tests

I had to disable exceptions in doctest, which downgrades our `REQUIRE`/`FAIL` macros to the `CHECK` level (which marks the test as failed but doesn't rage quit). We need to see if that impacts our test suites negatively, and if we should clarify this behavior by changing our existing `REQUIRE`/`FAIL` uses to `CHECK`.

This is work for a follow-up PR where I'd welcome a volunteer - for this PR we should just ensure that the tests still work properly.

## Further work

`godot-cpp` could likely reuse the same logic to build without extensions by default, with a command line option to override that behavior for users who want to compile with a library that needs exceptions (or their own code).

`3.x` backport: #80622