@akien-mga 2023-08-14
> I had to disable exceptions in doctest, which downgrades our `REQUIRE`/`FAIL` macros to the `CHECK` level (which marks the test as failed but doesn't rage quit). We need to see if that impacts our test suites negatively, and if we should clarify this behavior by changing our existing `REQUIRE`/`FAIL` uses to `CHECK`.

We seem to have one failing test suite after these changes:

```
$ ./bin/godot.linuxbsd.editor.dev.x86_64 --test
[doctest] doctest version is "2.4.11"
[doctest] run with "--help" for options
ERROR: Condition "singleton != nullptr" is true.
   at: NavigationServer3D (./servers/navigation_server_3d.cpp:189)
===============================================================================
./tests/core/io/test_image.h:221:
TEST CASE:  [Image] Modifying pixels of an image

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

./tests/core/io/test_image.h:271: ERROR: CHECK( img->get_pixel(x, y).is_equal_approx(Color(1, 1, 1, 1)) ) is NOT correct!
  values: CHECK( false )
  logged: fill_rect() should colorize all image pixels within rect bounds.

ERROR: Resource was not pre cached for the resource section, bug?
   at: _write_resource (./scene/resources/resource_format_text.cpp:1848)
WARNING: Couldn't load resource (no cache): local://Resource_7urp5
     at: parse_variant (./core/io/resource_format_binary.cpp:411)
ERROR: The object does not have any 'meta' values with the key 'next'.
   at: get_meta (./core/object/object.cpp:926)
ERROR: The object does not have any 'meta' values with the key 'next'.
   at: get_meta (./core/object/object.cpp:926)
```