@akien-mga 2023-08-14
We were already defining `-fno-exceptions` for Android, iOS and Web, so I removed those since it's now done in `SConstruct` directly.

Also removed the `ios_exceptions` option since it's redundant with `disable_exceptions` (as long as `-fexceptions` is the default behavior with Apple Clang, to be confirmed @bruvzg).