@Jordyfel 2023-08-14
Ran the production builds on my machine (6 core CPU), with MSVC, same scons options.

Editor build time goes from 12:52 to 7:36, a reduction of 41%.

Release template build time goes from 7:46 to 5:45, a reduction of 26%.