@Calinou 2023-08-14
**OS:** Fedora 38
**CPU:** Intel Core i9-13900K
**SSD:** Solidigm P44 Pro 2 TB
**Compiler:** GCC 13.2.1

<details>
<summary>Build options</summary>

Editor build options: `scons -j32 platform=linuxbsd linker=mold optimize=none fast_unsafe=yes progress=no debug_symbols=yes builtin_embree=no builtin_enet=no builtin_freetype=no builtin_graphite=no builtin_harfbuzz=no builtin_libogg=no builtin_libpng=no builtin_libtheora=no builtin_libvorbis=no builtin_libwebp=no builtin_mbedtls=no builtin_miniupnpc=no builtin_pcre2=no builtin_zlib=no builtin_zstd=no scu_build=all tests=no`

Release export template build options: `scons -j32 platform=linuxbsd target=template_release debug_symbols=yes lto=full progress=no scu_build=all`
</details>
I ran each build twice with `ccache -C` first to fully empty ccache.
```bash
# Before
vstask "Build (editor)"  1387.70s user 115.27s system 2425% cpu 1:01.96 total
Stripped binary size: 170,679,344 bytes

# After
vstask "Build (editor)"  1278.58s user 112.49s system 2425% cpu 57.361 total
Stripped binary size: 150,215,728 bytes
```

```bash
# Before
vstask "Build (release export template)"  2962.39s user 139.11s system 1888% cpu 2:44.21 total
Striped binary size: 70,403,424 bytes

# After
vstask "Build (release export template)"  2351.15s user 132.09s system 1904% cpu 2:10.37 total
Stripped binary size: 58,283,360 bytes
```